feature: remodel image upload mechanism
